feat: implement bulk mutation handling and reporting
- Added BulkMutation module with classes for handling job IDs, skipped jobs, and report building.
- Refactored dead letter support to utilize the new BulkMutation classes for better organization and code reuse.
- Updated in-memory queue store to integrate bulk mutation reporting for job retries and cancellations.
- Introduced RBS definitions for the new BulkMutation classes to enhance type safety and documentation.
