+++ core/karya/lib/karya/queue_store/in_memory.rb
@@ -11,6 +11,7 @@ require 'bigdecimal'
 require_relative 'base'
 require_relative 'bulk_mutation_report'
 require_relative '../circuit_breaker'
+require_relative '../internal/bulk_mutation'
 require_relative '../internal/failure_classification'
 require_relative '../internal/retry_policy_normalizer'
 require_relative 'queue_control_result'
